+++ examples/anyscale-v2-kitchensink/outputs.tf
@@ -11,8 +11,8 @@ output "registration_command" {
     --cloud-storage-bucket-name ${module.google_anyscale_v2_kitchensink.cloudstorage_bucket_name} \
     --filestore-instance-id ${module.google_anyscale_v2_kitchensink.filestore_name} \
     --filestore-location ${module.google_anyscale_v2_kitchensink.filestore_location} \
-    --anyscale-service-account-email ${module.google_anyscale_v2_kitchensink.iam_anyscale_access_role_email} \
-    --instance-service-account-email ${module.google_anyscale_v2_kitchensink.iam_anyscale_cluster_node_role_email} \
+    --anyscale-service-account-email ${module.google_anyscale_v2_kitchensink.iam_anyscale_access_service_acct_email} \
+    --instance-service-account-email ${module.google_anyscale_v2_kitchensink.iam_anyscale_cluster_node_service_acct_email} \
     --provider-name ${module.google_anyscale_v2_kitchensink.iam_workload_identity_provider_name} \
     --memorystore-instance-name ${module.google_anyscale_v2_kitchensink.memorystore_id}
   EOT