+++ modules/google-anyscale-iam/outputs.tf
@@ -1,24 +1,24 @@
-# -------------------------------------------
-# Anyscale Cross Account Access Role
-# -------------------------------------------
-output "iam_anyscale_access_role_id" {
-  description = "Anyscale cross account access role ID"
-  value       = try(google_service_account.anyscale_access_role[0].id, null)
+# ----------------------------------------------
+# Anyscale Cross Account Access Service Account
+# ----------------------------------------------
+output "iam_anyscale_access_service_acct_id" {
+  description = "Anyscale cross account access service account ID"
+  value       = try(google_service_account.anyscale_access_service_acct[0].id, null)
 }
 
-output "iam_anyscale_access_role_email" {
-  description = "Anyscale cross account access role email"
-  value       = try(google_service_account.anyscale_access_role[0].email, null)
+output "iam_anyscale_access_service_acct_email" {
+  description = "Anyscale cross account access service account email"
+  value       = try(google_service_account.anyscale_access_service_acct[0].email, null)
 }
 
-output "iam_anyscale_access_role_name" {
-  description = "Anyscale cross account access role name"
-  value       = try(google_service_account.anyscale_access_role[0].name, null)
+output "iam_anyscale_access_service_acct_name" {
+  description = "Anyscale cross account access service account name"
+  value       = try(google_service_account.anyscale_access_service_acct[0].name, null)
 }
 
-output "iam_anyscale_access_role_unique_id" {
-  description = "Anyscale cross account access role unique ID"
-  value       = try(google_service_account.anyscale_access_role[0].unique_id, null)
+output "iam_anyscale_access_service_acct_unique_id" {
+  description = "Anyscale cross account access service account unique ID"
+  value       = try(google_service_account.anyscale_access_service_acct[0].unique_id, null)
 }
 
 # -------------------------------------------
@@ -47,24 +47,24 @@ output "iam_workload_identity_provider_name" {
 }
 
 # -------------------------------------------
-# Anyscale Cluster Node Role
+# Anyscale Cluster Node Service Account
 # -------------------------------------------
-output "iam_anyscale_cluster_node_role_id" {
-  description = "Anyscale cross account cluster node role ID"
-  value       = try(google_service_account.anyscale_cluster_node_role[0].id, null)
+output "iam_anyscale_cluster_node_service_acct_id" {
+  description = "Anyscale cross account cluster node service account ID"
+  value       = try(google_service_account.anyscale_cluster_node_service_acct[0].id, null)
 }
 
-output "iam_anyscale_cluster_node_role_email" {
-  description = "Anyscale cross account cluster node role email"
-  value       = try(google_service_account.anyscale_cluster_node_role[0].email, null)
+output "iam_anyscale_cluster_node_service_acct_email" {
+  description = "Anyscale cross account cluster node service account email"
+  value       = try(google_service_account.anyscale_cluster_node_service_acct[0].email, null)
 }
 
-output "iam_anyscale_cluster_node_role_name" {
-  description = "Anyscale cross account cluster node role name"
-  value       = try(google_service_account.anyscale_cluster_node_role[0].name, null)
+output "iam_anyscale_cluster_node_service_acct_name" {
+  description = "Anyscale cross account cluster node service account name"
+  value       = try(google_service_account.anyscale_cluster_node_service_acct[0].name, null)
 }
 
-output "iam_anyscale_cluster_node_role_unique_id" {
-  description = "Anyscale cross account cluster node role unique ID"
-  value       = try(google_service_account.anyscale_cluster_node_role[0].unique_id, null)
+output "iam_anyscale_cluster_node_service_acct_unique_id" {
+  description = "Anyscale cross account cluster node service account unique ID"
+  value       = try(google_service_account.anyscale_cluster_node_service_acct[0].unique_id, null)
 }